music_player/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
